refactor: make record values part of api
