+++ engine/src/main/java/io/camunda/zeebe/streamplatform/api/RecordValues.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.streamplatform.api;
 
-import static io.camunda.zeebe.engine.processing.streamprocessor.TypedEventRegistry.EVENT_REGISTRY;
+import static io.camunda.zeebe.streamplatform.api.TypedEventRegistry.EVENT_REGISTRY;
 
 import io.camunda.zeebe.logstreams.log.LoggedEvent;
 import io.camunda.zeebe.protocol.impl.record.UnifiedRecordValue;
