+++ engine/src/main/java/io/camunda/zeebe/streamplatform/impl/ProcessingStateMachine.java
@@ -10,7 +10,6 @@ package io.camunda.zeebe.streamplatform.impl;
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDbTransaction;
 import io.camunda.zeebe.engine.metrics.StreamProcessorMetrics;
-import io.camunda.zeebe.engine.processing.streamprocessor.RecordValues;
 import io.camunda.zeebe.logstreams.impl.Loggers;
 import io.camunda.zeebe.logstreams.log.LogStreamBatchWriter;
 import io.camunda.zeebe.logstreams.log.LogStreamReader;
@@ -27,6 +26,7 @@ import io.camunda.zeebe.streamplatform.api.EmptyProcessingResult;
 import io.camunda.zeebe.streamplatform.api.ProcessingResult;
 import io.camunda.zeebe.streamplatform.api.ProcessingResultBuilder;
 import io.camunda.zeebe.streamplatform.api.RecordProcessor;
+import io.camunda.zeebe.streamplatform.api.RecordValues;
 import io.camunda.zeebe.streamplatform.api.TypedRecord;
 import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
 import io.camunda.zeebe.util.buffer.BufferUtil;
