+++ engine/src/main/java/io/camunda/zeebe/streamplatform/impl/ReplayStateMachine.java
@@ -10,7 +10,6 @@ package io.camunda.zeebe.streamplatform.impl;
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDbTransaction;
 import io.camunda.zeebe.engine.metrics.ReplayMetrics;
-import io.camunda.zeebe.engine.processing.streamprocessor.RecordValues;
 import io.camunda.zeebe.logstreams.impl.Loggers;
 import io.camunda.zeebe.logstreams.impl.log.LogStreamBatchReaderImpl;
 import io.camunda.zeebe.logstreams.log.LogRecordAwaiter;
@@ -28,6 +27,7 @@ import io.camunda.zeebe.scheduler.future.CompletableActorFuture;
 import io.camunda.zeebe.scheduler.retry.RecoverableRetryStrategy;
 import io.camunda.zeebe.scheduler.retry.RetryStrategy;
 import io.camunda.zeebe.streamplatform.api.RecordProcessor;
+import io.camunda.zeebe.streamplatform.api.RecordValues;
 import io.camunda.zeebe.streamplatform.api.TypedRecord;
 import io.camunda.zeebe.streamplatform.api.state.KeyGeneratorControls;
 import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
