+++ engine/src/main/java/io/camunda/zeebe/streamplatform/impl/StreamProcessorContext.java
@@ -8,7 +8,6 @@
 package io.camunda.zeebe.streamplatform.impl;
 
 import io.camunda.zeebe.db.TransactionContext;
-import io.camunda.zeebe.engine.processing.streamprocessor.RecordValues;
 import io.camunda.zeebe.logstreams.log.LogStream;
 import io.camunda.zeebe.logstreams.log.LogStreamBatchWriter;
 import io.camunda.zeebe.logstreams.log.LogStreamReader;
@@ -18,6 +17,7 @@ import io.camunda.zeebe.streamplatform.api.CommandResponseWriter;
 import io.camunda.zeebe.streamplatform.api.InterPartitionCommandSender;
 import io.camunda.zeebe.streamplatform.api.ProcessingScheduleService;
 import io.camunda.zeebe.streamplatform.api.ReadonlyStreamProcessorContext;
+import io.camunda.zeebe.streamplatform.api.RecordValues;
 import io.camunda.zeebe.streamplatform.api.TypedRecord;
 import io.camunda.zeebe.streamplatform.api.state.KeyGeneratorControls;
 import io.camunda.zeebe.streamplatform.api.state.MutableLastProcessedPositionState;
