+++ engine/src/test/java/io/camunda/zeebe/engine/util/TestStreams.java
@@ -18,7 +18,6 @@ import io.camunda.zeebe.db.ZeebeDb;
 import io.camunda.zeebe.db.ZeebeDbFactory;
 import io.camunda.zeebe.engine.Engine;
 import io.camunda.zeebe.engine.Loggers;
-import io.camunda.zeebe.engine.processing.streamprocessor.TypedEventRegistry;
 import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecordProcessorFactory;
 import io.camunda.zeebe.engine.state.EventApplier;
 import io.camunda.zeebe.engine.state.appliers.EventAppliers;
@@ -43,6 +42,7 @@ import io.camunda.zeebe.streamplatform.api.CommandResponseWriter;
 import io.camunda.zeebe.streamplatform.api.InterPartitionCommandSender;
 import io.camunda.zeebe.streamplatform.api.ReadonlyStreamProcessorContext;
 import io.camunda.zeebe.streamplatform.api.StreamProcessorLifecycleAware;
+import io.camunda.zeebe.streamplatform.api.TypedEventRegistry;
 import io.camunda.zeebe.streamplatform.impl.StreamProcessor;
 import io.camunda.zeebe.streamplatform.impl.StreamProcessorListener;
 import io.camunda.zeebe.streamplatform.impl.StreamProcessorMode;
